Remove absolute positioning of footer
Problem with absolute position is that height of footer needs to be defined => content can wrap outside footer.

Removing absolute position works fine on pages with content, but when few content is available (e.g. on search) then footer can appear above bottom of window. Setting a min-height to container of 100vh (= viewport height) solves this. And if we substract the height of the navbar, then the footer will appear just after the fold (i.e. just after you scroll down).